Backend/Enterprise_Database/requirements.txt
asgiref==3.8.1
captcha==0.6.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.4
django-cors-headers==4.6.0
django-recaptcha==4.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
numpy==2.2.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.3.0
pyotp==2.9.0